# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pitbull Suplementos API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  flavor?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountBadge?: string | null;
  stock: number;
  images: string[];
  featured: boolean;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  category: string;
  flavor?: string;
  price: number;
  originalPrice?: number;
  discountBadge?: string;
  stock: number;
  images?: string[];
  featured?: boolean;
  active?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  category?: string;
  flavor?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountBadge?: string | null;
  stock?: number;
  images?: string[];
  featured?: boolean;
  active?: boolean;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  token: string;
}

export interface AdminCheckResult {
  authenticated: boolean;
}

export type ListProductsParams = {
category?: string;
flavor?: string;
search?: string;
sort?: ListProductsSort;
featured?: boolean;
active?: boolean;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  name_asc: 'name_asc',
  featured: 'featured',
} as const;