# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pitbull Suplementos API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: admin
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with optional filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: flavor
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [price_asc, price_desc, name_asc, featured] }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Wrong password

  /admin/check:
    get:
      operationId: adminCheck
      tags: [admin]
      summary: Check admin auth status
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCheckResult"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Product:
      type: object
      required: [id, name, category, price, stock, images, featured, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        flavor:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountBadge:
          type: ["string", "null"]
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string

    ProductInput:
      type: object
      required: [name, category, price, stock]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        flavor:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        discountBadge:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        active:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        flavor:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountBadge:
          type: ["string", "null"]
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        active:
          type: boolean

    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminLoginResult:
      type: object
      required: [token]
      properties:
        token:
          type: string

    AdminCheckResult:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean