# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pitbull Suplementos API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  flavor?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountBadge?: string | null;
  stock: number;
  images: string[];
  featured: boolean;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}