# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pitbull Suplementos API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  category: string;
  flavor?: string;
  price: number;
  originalPrice?: number;
  discountBadge?: string;
  stock: number;
  images?: string[];
  featured?: boolean;
  active?: boolean;
}