# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pitbull Suplementos API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  description?: string;
  category?: string;
  flavor?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountBadge?: string | null;
  stock?: number;
  images?: string[];
  featured?: boolean;
  active?: boolean;
}